38 think about database optimization (#39)

+++ src/main/java/scari/corp/taro/dto/taro/CardResponseDto.java
@@ -19,7 +19,7 @@ public record CardResponseDto(
         String nameRu,
         String arcana,
         String suit,
-        Integer number,
+        Short number,
         boolean isReversed,
         String interpretation,
         String imageKey
